move exported test functions
